Extract functions to a utility folder
